+++ assets/plugins/wwsu-state/js/wwsu-state.js
@@ -1197,6 +1197,7 @@ class WWSUstate extends WWSUevents {
 					},
 					sport: {
 						type: "select",
+						hideNone: false,
 						validator: function (callback) {
 							let value = this.getValue();
 							if (!title || title !== value) {
@@ -1462,6 +1463,7 @@ class WWSUstate extends WWSUevents {
 					},
 					sport: {
 						type: "select",
+						hideNone: false,
 						validator: function (callback) {
 							let value = this.getValue();
 							if (!title || title !== value) {